refactor(Navbar): use jsx for components and js for containers
